clean up TimeOption models
