+++ src/models/index.ts
@@ -4,4 +4,5 @@ export * from './Preview';
 export * from './SortCategory';
 export * from './SortOptions';
 export * from './TimeCategory';
+export * from './TimeOptions';
 export * from './UpvoteDirection';
